% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{c.sam}
\alias{c.sam}
\title{Combine samjr fits into a \code{samset}}
\usage{
\method{c}{sam}(...)
}
\arguments{
\item{...}{one or more \code{sam} fits.}
}
\value{
a list of fits with class \code{samset}.
}
\description{
Combine samjr fits into a \code{samset}
}
