% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{catchplot}
\alias{catchplot}
\alias{catchplot.sam}
\alias{catchplot.samforecast}
\alias{catchplot.samset}
\title{samjr total catch plot}
\usage{
catchplot(fit, obs.show = TRUE, ...)

\method{catchplot}{sam}(fit, obs.show = TRUE, ...)

\method{catchplot}{samforecast}(fit, obs.show = TRUE, ...)

\method{catchplot}{samset}(fit, obs.show = TRUE, ...)
}
\arguments{
\item{fit}{object returned from \code{\link{sam.fit}}.}

\item{obs.show}{if \code{TRUE} (default) overlay the observed
sum-of-products catch in weight as crosses.}

\item{...}{extra arguments passed to \code{\link{plotit}}.}
}
\value{
invisibly a list returned by \code{plotit}.
}
\description{
Plot of estimated total catch in weight with two-standard-deviation
confidence band, optionally overlaid with the sum-of-products of the
observed catch in numbers and the catch mean weight.
}
