% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{catchtable}
\alias{catchtable}
\alias{catchtable.sam}
\title{Time series of estimated total catch in weight}
\usage{
catchtable(fit, ...)

\method{catchtable}{sam}(fit, ...)
}
\arguments{
\item{fit}{a fitted \code{sam} object.}

\item{...}{unused.}
}
\value{
matrix with Estimate, Low and High columns and years as row names.
}
\description{
Time series of estimated total catch in weight
}
