% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{caytable}
\alias{caytable}
\title{Catch-at-age in numbers from a samjr fit}
\usage{
caytable(fit, fleet = which(fit$data$fleetTypes == 0))
}
\arguments{
\item{fit}{a fitted \code{sam} object.}

\item{fleet}{integer; which catch fleet(s) to include (default = all
residual-catch fleets).}
}
\value{
matrix with year row names and age column names.
}
\description{
Catch-at-age in numbers from a samjr fit
}
