% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ckmrl.R
\name{ckmrltable}
\alias{ckmrltable}
\alias{ckmrltable.sam}
\title{Observed and expected length-based close-kin pair counts}
\usage{
ckmrltable(fit, ...)

\method{ckmrltable}{sam}(fit, ...)
}
\arguments{
\item{fit}{a fitted \code{sam} object.}

\item{...}{unused.}
}
\value{
the pair table with \code{ePOP}, \code{rPOP}, \code{eHSP},
\code{rHSP} added.
}
\description{
Observed and expected length-based close-kin pair counts
}
