% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{coef.sam}
\alias{coef.sam}
\title{Extract fixed-effect coefficients from a samjr fit}
\usage{
\method{coef}{sam}(object, ...)
}
\arguments{
\item{object}{a fitted \code{sam} object.}

\item{...}{unused.}
}
\value{
a named numeric vector with attributes \code{cov} (the
covariance) and \code{sd} (standard errors), class \code{samcoef}.
}
\description{
Extract fixed-effect coefficients from a samjr fit
}
