% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{faytable}
\alias{faytable}
\alias{faytable.sam}
\title{F-at-age table from a samjr fit}
\usage{
faytable(fit, ...)

\method{faytable}{sam}(fit, fleet = which(fit$data$fleetTypes == 0), ...)
}
\arguments{
\item{fit}{a fitted \code{sam} object.}

\item{...}{unused.}

\item{fleet}{integer; which catch fleet(s) to include (default = all
residual-catch fleets).}
}
\value{
matrix with year row names and age column names.
}
\description{
F-at-age table from a samjr fit
}
