% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{fbarplot}
\alias{fbarplot}
\alias{fbarplot.sam}
\alias{fbarplot.samforecast}
\alias{fbarplot.samset}
\title{samjr Fbar plot}
\usage{
fbarplot(fit, ...)

\method{fbarplot}{sam}(fit, ...)

\method{fbarplot}{samforecast}(fit, ...)

\method{fbarplot}{samset}(fit, ...)
}
\arguments{
\item{fit}{object returned from \code{\link{sam.fit}}.}

\item{...}{extra arguments passed to \code{\link{plotit}}.}
}
\value{
invisibly a list returned by \code{plotit}.
}
\description{
Plot of estimated average fishing mortality (\eqn{\bar F}) over the
age range given by \code{fit$conf$fbarRange}, with two-standard-deviation
confidence band.
}
