% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{fbartable}
\alias{fbartable}
\alias{fbartable.sam}
\title{Time series of estimated mean fishing mortality}
\usage{
fbartable(fit, ...)

\method{fbartable}{sam}(fit, ...)
}
\arguments{
\item{fit}{a fitted \code{sam} object.}

\item{...}{unused.}
}
\value{
matrix with Estimate, Low and High columns and years as row names.
}
\description{
Time series of estimated mean fishing mortality
}
