% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hcr.R
\name{icesAdviceRule}
\alias{icesAdviceRule}
\alias{icesAdviceRule.sam}
\title{ICES advice rule (F = 0 below Blim, ramp to Fmsy at MSYBtrigger)}
\usage{
icesAdviceRule(fit, ...)

\method{icesAdviceRule}{sam}(fit, Fmsy, MSYBtrigger, Blim, ...)
}
\arguments{
\item{fit}{a samjr fit.}

\item{...}{other arguments to \code{\link{hcr.sam}}.}

\item{Fmsy}{target F.}

\item{MSYBtrigger}{SSB above which the target is fully applied.}

\item{Blim}{below this SSB, F is set to zero.}
}
\description{
Thin wrapper around \code{\link{hcr}}.
}
\examples{
\donttest{
data(nscodData); data(nscodConf)
fit <- samjr::sam.fit(nscodData, nscodConf,
                      samjr::defpar(nscodData, nscodConf), silent = TRUE)
fit$conf$stockRecruitmentModelCode <- 2L
fitBH <- samjr::runwithout(fit)
hh <- icesAdviceRule(fitBH, Fmsy = 0.22, MSYBtrigger = 150000,
                     Blim = 100000, nosim = 100, nYears = 5, seed = 1)
hh$projection$SSB
}
}
