% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retro.R
\name{leaveout}
\alias{leaveout}
\title{Leave-one-fleet-out analysis of a samjr fit}
\usage{
leaveout(fit, fleet = as.list(2:fit$data$noFleets), ncores = 1, ...)
}
\arguments{
\item{fit}{a fitted \code{sam} object.}

\item{fleet}{a list of fleet vectors; one run per element.}

\item{ncores}{ignored in samjr.}

\item{...}{extra arguments forwarded to \code{\link{sam.fit}}.}
}
\value{
a \code{samset}.
}
\description{
Leave-one-fleet-out analysis of a samjr fit
}
