% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{nobs.sam}
\alias{nobs.sam}
\title{Number of observations in a samjr fit}
\usage{
\method{nobs}{sam}(object, ...)
}
\arguments{
\item{object}{a fitted \code{sam} object from \code{\link{sam.fit}}.}

\item{...}{extra arguments (ignored).}
}
\value{
an integer.
}
\description{
Number of observations in a samjr fit
}
