% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{partable}
\alias{partable}
\alias{partable.sam}
\title{Parameter table from a samjr fit}
\usage{
partable(fit, ...)

\method{partable}{sam}(fit, ...)
}
\arguments{
\item{fit}{a fitted \code{sam} object.}

\item{...}{unused.}
}
\value{
matrix with par, sd(par), exp(par), Low, High columns.
}
\description{
Parameter table from a samjr fit
}
