% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.sam}
\alias{plot.sam}
\title{Default 2x2 panel plot for a samjr fit}
\usage{
\method{plot}{sam}(x, ...)
}
\arguments{
\item{x}{a \code{sam} object returned by \code{\link{sam.fit}}.}

\item{...}{extra arguments passed to each panel.}
}
\value{
invisible \code{NULL}.
}
\description{
Produces a 2-by-2 layout showing SSB, Fbar, recruitment and catch.
}
