% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.samforecast}
\alias{plot.samforecast}
\title{Default 2x2 panel plot for a samjr forecast}
\usage{
\method{plot}{samforecast}(x, ...)
}
\arguments{
\item{x}{a \code{samforecast} object returned by \code{\link{forecast}}.}

\item{...}{extra arguments passed to each panel.}
}
\value{
invisible \code{NULL}.
}
\description{
Default 2x2 panel plot for a samjr forecast
}
