% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hcr.R
\name{plot.samref_hcr}
\alias{plot.samref_hcr}
\title{2x2 overview plot of an \code{samref_hcr} object: SSB, F-bar,
Recruitment, Catch.}
\usage{
\method{plot}{samref_hcr}(x, ...)
}
\arguments{
\item{x}{a \code{samref_hcr} object.}

\item{...}{extra arguments passed to the panel plots.}
}
\value{
invisible \code{NULL}.
}
\description{
2x2 overview plot of an \code{samref_hcr} object: SSB, F-bar,
Recruitment, Catch.
}
\examples{
\donttest{
data(nscodData); data(nscodConf)
fit <- samjr::sam.fit(nscodData, nscodConf,
                      samjr::defpar(nscodData, nscodConf), silent = TRUE)
fit$conf$stockRecruitmentModelCode <- 2L
fitBH <- samjr::runwithout(fit)
hh <- icesAdviceRule(fitBH, Fmsy = 0.22, MSYBtrigger = 150000,
                     Blim = 100000, nosim = 200, nYears = 5, seed = 1)
plot(hh)         # 2x2 overview
ssbplot(hh)      # SSB only with rule lines
fbarplot(hh)
}
}
