% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.samset}
\alias{plot.samset}
\title{Default 2x2 panel plot for a samjr samset}
\usage{
\method{plot}{samset}(x, ...)
}
\arguments{
\item{x}{a \code{samset} object (as returned by \code{\link{retro}},
\code{\link{leaveout}}, \code{\link{jit}} or \code{\link{c.sam}}).}

\item{...}{extra arguments passed to each panel.}
}
\value{
invisible \code{NULL}.
}
\description{
Default 2x2 panel plot for a samjr samset
}
