% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{plot.samypr}
\alias{plot.samypr}
\title{Plot a YPR curve}
\usage{
\method{plot}{samypr}(x, ...)
}
\arguments{
\item{x}{a \code{samypr} object from \code{\link{ypr}}.}

\item{...}{extra arguments passed to \code{plot}.}
}
\value{
invisible \code{NULL}.
}
\description{
Plot a YPR curve
}
