% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recruitment.R
\name{predictLogR}
\alias{predictLogR}
\title{Predicted log-recruitment for an SR model}
\usage{
predictLogR(srCode, recPars, logSsb)
}
\arguments{
\item{srCode}{1 (Ricker) or 2 (Beverton-Holt).}

\item{recPars}{length-2 vector.}

\item{logSsb}{scalar or vector log-SSB.}
}
\value{
scalar or vector log-recruitment.
}
\description{
Closed-form predictor matching the closures in
\code{samjr/R/likelihood.R}. Useful for SR-curve overlays.
}
\examples{
## Ricker R(S) at SSB = exp(10..12)
predictLogR(1L, c(1.5, -3), logSsb = 10:12)
}
