% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.samcoef}
\alias{print.samcoef}
\title{Print method for \code{samcoef}}
\usage{
\method{print}{samcoef}(x, ...)
}
\arguments{
\item{x}{a \code{samcoef} object.}

\item{...}{unused.}
}
\description{
Print method for \code{samcoef}
}
