% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast.R
\name{print.samforecast}
\alias{print.samforecast}
\title{Print a samjr forecast}
\usage{
\method{print}{samforecast}(x, ...)
}
\arguments{
\item{x}{a \code{samforecast} object.}

\item{...}{unused.}
}
\description{
Print a samjr forecast
}
