% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{print.samqtable}
\alias{print.samqtable}
\title{Print method for \code{samqtable}}
\usage{
\method{print}{samqtable}(x, ...)
}
\arguments{
\item{x}{a \code{samqtable}.}

\item{...}{unused.}
}
\description{
Print method for \code{samqtable}
}
