% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.samres}
\alias{print.samres}
\title{Print a \code{samres} residual data frame}
\usage{
\method{print}{samres}(x, ...)
}
\arguments{
\item{x}{a \code{samres} object.}

\item{...}{unused.}
}
\description{
Print a \code{samres} residual data frame
}
