% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.samset}
\alias{print.samset}
\title{Print a \code{samset}}
\usage{
\method{print}{samset}(x, ...)
}
\arguments{
\item{x}{a \code{samset} (list of \code{sam} fits).}

\item{...}{unused.}
}
\description{
Print a \code{samset}
}
