% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{print.samypr}
\alias{print.samypr}
\title{Print a YPR summary table}
\usage{
\method{print}{samypr}(x, ...)
}
\arguments{
\item{x}{a \code{samypr} object from \code{\link{ypr}}.}

\item{...}{unused.}
}
\description{
Print a YPR summary table
}
