% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recruitment.R
\name{recPars}
\alias{recPars}
\title{Extract SR parameters from a samjr fit}
\usage{
recPars(fit)
}
\arguments{
\item{fit}{a samjr \code{sam} object.}
}
\value{
length-2 numeric vector of SR parameters or
  \code{numeric(0)} for SR code 0.
}
\description{
Extract SR parameters from a samjr fit
}
\examples{
\donttest{
data(nscodData); data(nscodConf)
fit <- samjr::sam.fit(nscodData, nscodConf,
                      samjr::defpar(nscodData, nscodConf), silent = TRUE)
recPars(fit)   # numeric(0) for random-walk recruitment
}
}
