% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{recplot}
\alias{recplot}
\alias{recplot.sam}
\alias{recplot.samforecast}
\alias{recplot.samset}
\title{samjr recruitment plot}
\usage{
recplot(fit, ...)

\method{recplot}{sam}(fit, ...)

\method{recplot}{samforecast}(fit, ...)

\method{recplot}{samset}(fit, ...)
}
\arguments{
\item{fit}{object returned from \code{\link{sam.fit}}.}

\item{...}{extra arguments passed to \code{\link{plotit}}.}
}
\value{
invisibly a list returned by \code{plotit}.
}
\description{
Plot of estimated recruitment (numbers in the lowest age class) with
two-standard-deviation confidence band.
}
