% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{rectable}
\alias{rectable}
\alias{rectable.sam}
\title{Time series of estimated recruitment}
\usage{
rectable(fit, ...)

\method{rectable}{sam}(fit, ...)
}
\arguments{
\item{fit}{a fitted \code{sam} object.}

\item{...}{unused.}
}
\value{
matrix with Estimate, Low and High columns and years as row names.
}
\description{
Time series of estimated recruitment
}
