% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retro.R
\name{retro}
\alias{retro}
\alias{retro.sam}
\title{Retrospective analysis of a samjr fit}
\usage{
retro(fit, year = NULL, ncores = 1, ...)

\method{retro}{sam}(fit, year = NULL, ncores = 1, ...)
}
\arguments{
\item{fit}{a fitted \code{sam} object.}

\item{year}{either a single integer \code{n} (peel 1..n years) or a
vector / matrix of years to drop (see SAM's \code{retro} for details).}

\item{ncores}{ignored in samjr (always serial).}

\item{...}{extra arguments forwarded to \code{\link{sam.fit}}.}
}
\value{
a list of fits with class \code{samset}.
}
\description{
Refits the assessment after peeling 1, 2, ..., \code{year} years
from the right edge of the time series. Returns a \code{samset}.
}
