% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samjr-package.R
\docType{package}
\name{samjr-package}
\alias{samjr-package}
\alias{samjr}
\title{samjr: a minimal RTMB re-implementation of SAM}
\description{
Provides a small subset of the SAM \code{stockassessment} user interface
(\code{\link{read.ices}}, \code{\link{setup.sam.data}}, \code{\link{defcon}},
\code{\link{defpar}}, \code{\link{sam.fit}}) backed by a pure-R RTMB
implementation of the state-space age-structured assessment model. Plotting
helpers (\code{\link{ssbplot}}, \code{\link{fbarplot}}, \code{\link{recplot}},
\code{\link{catchplot}}) mirror the corresponding SAM functions.
}
\details{
Also provides deterministic per-recruit and equilibrium reference points
(\code{\link{referencepoints}}, \code{\link{perRecruitClosure}}) with
delta-method confidence intervals, plus stochastic harvest control rule
projections (\code{\link{hcr}}, \code{\link{icesAdviceRule}}), mirroring the
reference-point catalogue of SAM's \code{stockassessment} package.
}
\author{
\strong{Maintainer}: Anders Nielsen \email{fishfollower@gmail.com}

Authors:
\itemize{
  \item Christoffer Moesgaard Albertsen \email{cmoe@aqua.dtu.dk}
}

}
\keyword{package}
