% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hcr.R
\name{samref_hcr-plots}
\alias{samref_hcr-plots}
\alias{ssbplot.samref_hcr}
\alias{fbarplot.samref_hcr}
\alias{recplot.samref_hcr}
\alias{catchplot.samref_hcr}
\title{Plot the SSB / F-bar / Recruitment / Catch trajectory of a samref HCR
projection.}
\usage{
\method{ssbplot}{samref_hcr}(fit, ...)

\method{fbarplot}{samref_hcr}(fit, ...)

\method{recplot}{samref_hcr}(fit, ...)

\method{catchplot}{samref_hcr}(fit, ...)
}
\arguments{
\item{fit}{a \code{samref_hcr} object as returned by \code{\link{hcr}}.}

\item{...}{extra graphical parameters passed to \code{plot}.}
}
\value{
invisible \code{NULL}.
}
\description{
Each method draws one year-by-year quantity with the 95\% ribbon and
(where relevant) dashed reference lines for the rule parameters.
Recruitment uses age \code{minAge}.
}
