% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{ssbplot}
\alias{ssbplot}
\alias{ssbplot.sam}
\alias{ssbplot.samforecast}
\alias{ssbplot.samset}
\title{samjr SSB plot}
\usage{
ssbplot(fit, ...)

\method{ssbplot}{sam}(fit, ...)

\method{ssbplot}{samforecast}(fit, ...)

\method{ssbplot}{samset}(fit, ...)
}
\arguments{
\item{fit}{object returned from \code{\link{sam.fit}}.}

\item{...}{extra arguments passed to \code{\link{plotit}}.}
}
\value{
invisibly a list returned by \code{plotit}.
}
\description{
Plot of estimated spawning stock biomass with two-standard-deviation
confidence band.
}
