% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{ssbtable}
\alias{ssbtable}
\alias{ssbtable.sam}
\title{Time series of estimated spawning stock biomass}
\usage{
ssbtable(fit, ...)

\method{ssbtable}{sam}(fit, ...)
}
\arguments{
\item{fit}{a fitted \code{sam} object.}

\item{...}{unused.}
}
\value{
matrix with Estimate, Low and High columns and years as row names.
}
\description{
Time series of estimated spawning stock biomass
}
