% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{tsbplot}
\alias{tsbplot}
\alias{tsbplot.sam}
\alias{tsbplot.samforecast}
\alias{tsbplot.samset}
\title{samjr total stock biomass plot}
\usage{
tsbplot(fit, ...)

\method{tsbplot}{sam}(fit, ...)

\method{tsbplot}{samforecast}(fit, ...)

\method{tsbplot}{samset}(fit, ...)
}
\arguments{
\item{fit}{object returned from \code{\link{sam.fit}}.}

\item{...}{extra arguments passed to \code{\link{plotit}}.}
}
\value{
invisibly a list returned by \code{plotit}.
}
\description{
Plot of estimated total stock biomass (\eqn{\sum_a N_{y,a} \cdot SW_{y,a}})
with two-standard-deviation confidence band.
}
