% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{tsbtable}
\alias{tsbtable}
\alias{tsbtable.sam}
\title{Time series of estimated total stock biomass}
\usage{
tsbtable(fit, ...)

\method{tsbtable}{sam}(fit, ...)
}
\arguments{
\item{fit}{a fitted \code{sam} object.}

\item{...}{unused.}
}
\value{
matrix with Estimate, Low and High columns.
}
\description{
Time series of estimated total stock biomass
}
