% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writing.R
\name{write.surveys}
\alias{write.surveys}
\title{Write the survey fleets of a samjr data object to an ICES survey file}
\usage{
write.surveys(data, fileout, ...)
}
\arguments{
\item{data}{a \code{sam_data} list or a fitted \code{sam} object.}

\item{fileout}{output file path.}

\item{...}{arguments forwarded to \code{\link{write}}.}
}
\value{
invisibly \code{NULL}.
}
\description{
Write the survey fleets of a samjr data object to an ICES survey file
}
